requirements.txt
Pillow>=5.3.0,<5.4.0
django-cors-headers>=3.7.0,<3.8.0
pymysql>=1.0.2,<1.1.0
boto3>=1.17.44,<1.18.0
django-storages>=1.11.1,<1.12.0
django-mptt==0.11.0
factory-boy==3.2.0
Faker==9.5.2
djangorestframework-recursive==0.1.2


   
appdirs==1.4.4
asgiref==3.3.1
click==8.0.0
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
django-leaflet==0.27.1
djangorestframework==3.12.2
djangorestframework-gis==0.17
flake8==3.9.0
# mccabe==0.6.1
mypy-extensions==0.4.3
# pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1
# toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.2
gunicorn>=19.9.0,<19.10.0

black==22.3.0
isort==5.10.1
requests==2.27.1